# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_parse_api_parse_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": ".xlsx, max 2 MB",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_parse_api_parse_post",
        "type": "object"
      },
...
      "CatalogCounts": {
        "properties": {
          "excluded": {
            "title": "Excluded",
            "type": "integer"
          },
          "excluded_by_reason": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Excluded By Reason",
            "type": "object"
          },
          "included": {
            "title": "Included",
            "type": "integer"
          },
          "included_by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Included By Source",
            "type": "object"
          }
        },
        "required": [
          "included",
          "included_by_source",
          "excluded",
          "excluded_by_reason"
        ],
        "title": "CatalogCounts",
        "type": "object"
      },
      "CatalogMeta": {
        "properties": {
          "age_days": {
            "title": "Age Days",
            "type": "integer"
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "counts": {
            "$ref": "#/components/schemas/CatalogCounts"
          },
          "licence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licence"
          },
          "snapshot_date": {
            "format": "date",
            "title": "Snapshot Date",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CatalogSource"
            },
            "title": "Sources",
            "type": "array"
          },
          "stale": {
            "description": "True when the snapshot is older than 3 days.",
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "snapshot_date",
          "age_days",
          "stale",
          "counts",
          "sources",
          "attribution"
        ],
        "title": "CatalogMeta",
        "type": "object"
      },
      "CatalogSource": {
        "properties": {
          "file_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Date"
          },
          "licence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licence"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name"
        ],
        "title": "CatalogSource",
        "type": "object"
      },
      "CompareFilters": {
        "additionalProperties": false,
        "properties": {
          "price_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "placet",
                  "mlibero"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "title": "CompareFilters",
        "type": "object"
      },
      "CompareRequest": {
        "additionalProperties": false,
        "description": "12 consecutive months of household consumption. Processed in memory\nand never stored or logged.",
        "properties": {
          "committed_power_kw": {
            "default": 3.0,
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Committed Power Kw",
            "type": "number"
          },
          "consumption": {
            "items": {
              "$ref": "#/components/schemas/MonthInput"
            },
            "maxItems": 12,
            "minItems": 12,
            "title": "Consumption",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/CompareFilters"
          },
          "istat_comune": {
            "anyOf": [
              {
                "pattern": "^\\d{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "6-digit ISTAT comune code; needed to see area-restricted offers.",
            "title": "Istat Comune"
          },
          "residency": {
            "default": "resident",
            "enum": [
              "resident",
              "non_resident"
            ],
            "title": "Residency",
            "type": "string"
          },
          "scenario": {
            "discriminator": {
              "mapping": {
                "flat": "#/components/schemas/Flat",
                "historical": "#/components/schemas/Historical",
                "scaled": "#/components/schemas/Scaled"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Historical"
              },
              {
                "$ref": "#/components/schemas/Scaled"
              },
              {
                "$ref": "#/components/schemas/Flat"
              }
            ],
            "title": "Scenario"
          },
          "top_n": {
            "default": 50,
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Top N",
            "type": "integer"
          }
        },
        "required": [
          "consumption"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
      "CompareResponse": {
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/Assumptions"
          },
          "excluded_by_reason": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Excluded By Reason",
            "type": "object"
          },
          "excluded_count": {
            "title": "Excluded Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ResultItem"
            },
            "title": "Results",
            "type": "array"
          },
          "snapshot_date": {
            "format": "date",
            "title": "Snapshot Date",
            "type": "string"
          },
          "total_eligible": {
            "description": "Offers eligible for this household.",
            "title": "Total Eligible",
            "type": "integer"
          },
          "total_matching": {
            "description": "Eligible offers matching the filters.",
            "title": "Total Matching",
            "type": "integer"
          }
        },
        "required": [
          "snapshot_date",
          "assumptions",
          "total_eligible",
          "total_matching",
          "results",
          "excluded_count",
          "excluded_by_reason"
        ],
        "title": "CompareResponse",
        "type": "object"
      },
      "ComuneOut": {
        "properties": {
          "codice": {
            "description": "6-digit ISTAT comune code.",
            "title": "Codice",
            "type": "string"
          },
          "nome": {
            "title": "Nome",
            "type": "string"
          },
          "regione": {
            "title": "Regione",
            "type": "string"
          },
          "sigla_provincia": {
            "title": "Sigla Provincia",
            "type": "string"
          }
        },
        "required": [
          "codice",
          "nome",
          "sigla_provincia",
          "regione"
        ],
        "title": "ComuneOut",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "Flat": {
        "additionalProperties": false,
        "description": "Replace every month's PUN with a single flat value.",
        "properties": {
          "kind": {
            "const": "flat",
            "default": "flat",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "minimum": 0.0,
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "value"
        ],
        "title": "Flat",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Historical": {
        "additionalProperties": false,
        "description": "The default scenario: use the PUN actually recorded, month by month.",
        "properties": {
          "kind": {
            "const": "historical",
            "default": "historical",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "Historical",
        "type": "object"
      },
      "MonthInput": {
        "additionalProperties": false,
        "properties": {
          "f1": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "F1"
          },
          "f2": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "F2"
          },
          "f3": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "F3"
          },
          "kwh": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Kwh",
            "type": "number"
          },
          "month": {
            "examples": [
              "2025-09"
            ],
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "title": "Month",
            "type": "string"
          }
        },
        "required": [
          "month",
          "kwh"
        ],
        "title": "MonthInput",
        "type": "object"
      },
...
      "OfferPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OfferListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "OfferPage",
        "type": "object"
      },
      "OfferSource": {
        "enum": [
          "placet",
          "mlibero",
          "custom"
        ],
        "title": "OfferSource",
        "type": "string"
      },
      "ParseResult": {
        "properties": {
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/ParsedProfile"
            },
            "title": "Profiles",
            "type": "array"
          },
          "skipped": {
            "description": "Locations found in the file but not usable (need 12 consecutive months of non-negative kWh).",
            "items": {
              "type": "string"
            },
            "title": "Skipped",
            "type": "array"
          }
        },
        "required": [
          "profiles",
          "skipped"
        ],
        "title": "ParseResult",
        "type": "object"
      },
      "ParsedProfile": {
        "properties": {
          "location": {
            "title": "Location",
            "type": "string"
          },
          "months": {
            "items": {
              "$ref": "#/components/schemas/SampleMonth"
            },
            "title": "Months",
            "type": "array"
          }
        },
        "required": [
          "location",
          "months"
        ],
        "title": "ParsedProfile",
        "type": "object"
      },
      "PriceType": {
        "enum": [
          "fixed",
          "variable"
        ],
        "title": "PriceType",
        "type": "string"
      },
...
      "SampleMonth": {
        "properties": {
          "kwh": {
            "title": "Kwh",
            "type": "number"
          },
          "month": {
            "examples": [
              "2025-09"
            ],
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "title": "Month",
            "type": "string"
          },
          "pun": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pun"
          }
        },
        "required": [
          "month",
          "kwh"
        ],
        "title": "SampleMonth",
        "type": "object"
      },
      "Scaled": {
        "additionalProperties": false,
        "description": "Multiply every historical PUN value by a factor (e.g. 1.2 = +20%).",
        "properties": {
          "factor": {
            "exclusiveMinimum": 0.0,
            "title": "Factor",
            "type": "number"
          },
          "kind": {
            "const": "scaled",
            "default": "scaled",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "factor"
        ],
        "title": "Scaled",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ValidationErrorResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Stateless API: consumption is processed in memory per request and never stored or logged. Every cost is the commodity/retailer cost only, before VAT ('costo materia energia (IVA esclusa)'), and a backtest ('perfect foresight'): the same kWh and the same monthly PUN as your last 12 months. It is not a forecast.",
    "title": "BestBill API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/catalog/meta": {
      "get": {
        "operationId": "catalog_meta_api_catalog_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalogue not loaded yet."
          }
        },
        "summary": "Catalog Meta"
      }
    },
    "/api/compare": {
      "post": {
        "description": "Backtest (\"perfect foresight\"): what today's offers would have cost on\nyour last 12 months of consumption and PUN. Commodity/retailer cost only,\nbefore VAT. Not a forecast.",
        "operationId": "compare_offers_api_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body over 64 KB."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalogue not loaded yet."
          }
        },
        "summary": "Compare Offers"
      }
    },
    "/api/comuni": {
      "get": {
        "operationId": "comuni_api_comuni_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 60,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ComuneOut"
                  },
                  "title": "Response Comuni Api Comuni Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comuni"
      }
    },
...
    "/api/offers": {
      "get": {
        "operationId": "list_offers_api_offers_get",
        "parameters": [
          {
            "in": "query",
            "name": "price_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PriceType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price Type"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OfferSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "supplier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Catalogue not loaded yet."
          }
        },
        "summary": "List Offers"
      }
    },
    "/api/parse": {
      "post": {
        "description": "Read a workbook in the legacy format. Nothing is stored.",
        "operationId": "parse_api_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_api_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upload over 2 MB."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Parse"
      }
    },
...
  }
}